+++ src/microgrid.rs
@@ -3,11 +3,22 @@
 
 //! High-level interface for the Microgrid API.
 
+mod bounds_aggregation;
+
 mod battery_bounds_tracker;
 mod battery_pool;
 pub use battery_pool::BatteryPool;
 
+mod pv_bounds_tracker;
+mod pv_pool;
+pub use pv_pool::PvPool;
+
 pub(crate) mod telemetry_tracker;
+pub use telemetry_tracker::battery_pool_telemetry_tracker::{
+    BatteryPoolSnapshot, BatteryPoolTelemetryTracker, InverterBatteryGroup,
+};
+pub use telemetry_tracker::inverter_battery_group_telemetry_tracker::InverterBatteryGroupStatus;
+pub use telemetry_tracker::pv_pool_telemetry_tracker::{PvPoolSnapshot, PvPoolTelemetryTracker};
 
 use crate::{Error, LogicalMeterConfig, LogicalMeterHandle, MicrogridClientHandle};
 
@@ -63,6 +74,14 @@ impl Microgrid {
 
     pub fn battery_pool(&self, component_ids: Option<Vec<u64>>) -> Result<BatteryPool, Error> {
         BatteryPool::try_new(
+            component_ids.map(|ids| ids.into_iter().collect()),
+            self.client.clone(),
+            self.logical_meter.clone(),
+        )
+    }
+
+    pub fn pv_pool(&self, component_ids: Option<Vec<u64>>) -> Result<PvPool, Error> {
+        PvPool::try_new(
             component_ids.map(|ids| ids.into_iter().collect()),
             self.client.clone(),
             self.logical_meter.clone(),